migration/src/requirements.txt
psycopg2==2.9.5
yoyo-migrations==8.2.0